# src/main/resources/person-rest-api_swagger.yaml
swagger: '2.0'
info:
  description: '"Spring Boot REST API Performing CRUD Operation"'
  version: 1.0.0
  title: REST API Performing CRUD Operation
host: 'localhost:9080'
basePath: /person-rest-api/v1
tags:
- name: person-controller
  description: Person Controller
paths:
  /persons:
    get:
      tags:
      - person-controller
      summary: getAllPersons
      operationId: getAllPersonsUsingGET
      produces:
      - '*/*'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Response'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
      - person-controller
      summary: createPerson
      operationId: createPersonUsingPOST
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - in: body
        name: request
        description: request
        required: true
        schema:
          $ref: '#/definitions/Request'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/Response'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/persons/{id}':
    get:
      tags:
      - person-controller
      summary: getPersonById
      operationId: getPersonByIdUsingGET
      produces:
      - '*/*'
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: integer
        format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Response'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
      - person-controller
      summary: updatePerson
      operationId: updatePersonUsingPUT
      consumes:
      - application/json
      produces:
      - '*/*'
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: integer
        format: int64
      - in: body
        name: request
        description: request
        required: true
        schema:
          $ref: '#/definitions/Request'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Response'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
      - person-controller
      summary: deletePerson
      operationId: deletePersonUsingDELETE
      produces:
      - '*/*'
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: integer
        format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  Person:
    type: object
    properties:
      age:
        type: string
      favourite_colour:
        type: string
      first_name:
        type: string
      id:
        type: integer
        format: int64
      last_name:
        type: string
    title: Person
  Request:
    type: object
    properties:
      person:
        type: array
        items:
          $ref: '#/definitions/Person'
    title: Request
  Response:
    type: object
    properties:
      person:
        type: array
        items:
          $ref: '#/definitions/Person'
    title: Response